updated centroids to float

+++ lab1_io.c
@@ -26,12 +26,12 @@ void clusters_out (const char* cluster_filename, int N, int* cluster_points){
 	fclose(fout);
 }
 
-void centroids_out (const char* centroid_filename, int K, int num_iterations, int* centroids){
+void centroids_out (const char* centroid_filename, int K, int num_iterations, float* centroids){
 	FILE *fout = fopen(centroid_filename, "w");
 
 	for (int i = 0; i < num_iterations+1; i++){				//ith iteration
 		for (int j = 0; j < K; j++){			//jth centroid of ith iteration
-			fprintf(fout, "%d %d %d, ", 
+			fprintf(fout, "%f %f %f, ", 
 									*(centroids+(i*K+j)*3), 	 //x coordinate
 									*(centroids+(i*K+j)*3+1),  //y coordinate
 									*(centroids+(i*K+j)*3+2)); //z coordinate
